% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_tools_mltp.R
\name{alignmentCanonicalMltp}
\alias{alignmentCanonicalMltp}
\alias{alignmentCovariationMltp}
\alias{helixCovariationMltp}
\alias{helixConservationMltp}
\alias{helixCanonicalMltp}
\alias{baseConservationMltp}
\title{Compute statistics for a multiple sequence alignments}
\usage{
alignmentCanonicalMltp(
  msa,
  helix,
  top.name = "FALSE",
  sort = TRUE,
  comp.seq = FALSE
)

alignmentCovariationMltp(
  msa,
  helix,
  top.name = "FALSE",
  sort = TRUE,
  comp.seq = FALSE
)

helixCovariationMltp(
  msa,
  helix,
  top.name = "FALSE",
  sort = TRUE,
  comp.seq = FALSE
)

helixConservationMltp(
  msa,
  helix,
  top.name = "FALSE",
  sort = TRUE,
  comp.seq = FALSE
)

helixCanonicalMltp(
  msa,
  helix,
  top.name = "FALSE",
  sort = TRUE,
  comp.seq = FALSE
)

baseConservationMltp(msa, pos, GroupName)
}
\arguments{
\item{msa}{Multiple sequence alignment as an array of named characters,
all of equal length. Typically output of readBstringSet.}

\item{helix}{Helix data.tables, with the 6 mandatory columns. "col" refer to a styling column, and will be used for styling the helix. See example for styling usage.}

\item{top.name}{Name of entity, which will be set as top name and will refer to first entity in plot}

\item{sort}{sort entities by alphabetic order}

\item{comp.seq}{Option to compare sequences from different entities by their "name" and provide
proper covariance colouring in future}
}
\description{
Functions to compute covariation, percent identity conservation,
and percent canonical basepairs given a multiple sequence alignment
and optionally a secondary structure. Statistics can be computed for a
single base, basepair, helix or entire alignment.
}
\details{
Conservation values have a range of [0, 1], where 0 is the absence of
primary sequence conservation (all bases different), and 1 is full primary
sequence conservation (all bases identical).

Canonical values have a range of [0, 1], where 0 is a complete lack of
basepair potential, and 1 indicates that all basepairs are valid

Covariation values have a range of [-2, 2], where -2 is a complete
lack of basepair potential and sequence conservation, 0 is complete
sequence conservation regardless of basepairing potential, and 2 is a
complete lack of sequence conservation but maintaining full basepair potential.

helix values are average of base/basepair values, and the alignment
values are averages of helices or all columns depending on whether
the helix argument is required.
}
