% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourByMltp.R
\name{colourByValueMltp}
\alias{colourByValueMltp}
\alias{colourByUnknottedGroupsMltp}
\alias{colourByConservationMltp}
\alias{colourByCanonicalMltp}
\alias{colourByCountMltp}
\alias{colourByCovariationMltp}
\alias{colourByBasepairFrequencyMltp}
\title{Assign colours to helices}
\usage{
colourByValueMltp(
  helix,
  cols,
  breaks,
  get = FALSE,
  log = FALSE,
  include.lowest = TRUE
)

colourByUnknottedGroupsMltp(helix, cols, get = FALSE)

colourByConservationMltp(
  helix,
  msa,
  cols,
  get = FALSE,
  top.name = "FALSE",
  sort = TRUE,
  comp.seq = FALSE
)

colourByCanonicalMltp(
  helix,
  msa,
  cols,
  get = FALSE,
  top.name = "FALSE",
  sort = TRUE,
  comp.seq = FALSE
)

colourByCountMltp(helix, cols, counts, get = FALSE)

colourByCovariationMltp(
  helix,
  msa,
  cols,
  get = FALSE,
  top.name = "FALSE",
  sort = TRUE,
  comp.seq = FALSE
)

colourByBasepairFrequencyMltp(helix, cols, get = TRUE)
}
\arguments{
\item{helix}{A helix data table to be coloured.}

\item{cols}{An array of characters (or numbers) representing a set of colours to
colour helix with. When missing, a default set of colours from defaultPalette()
will be used. Valid input include hex codes, colour names from the colours function,
and integer numbers. The colours will be interpreted as being from best to worst.}

\item{breaks}{An integer number of intervals to break the ‘value’ column of helix
into, or a list of numbers defining the interval breaks. If missing, the range of
‘helix$value’ will automatically be split evenly into intervals for each colour available.}

\item{get}{If TRUE, returns the input helix with a col column, else simply
returns an array of colours the same length as the number of row in helix.
The exceptions are colourByBasepairFrequency and colourByUnknottedGroupsMltp
which will return a different helix if TRUE, and a list of colours that
will not match the input helix if FALSE.}

\item{log}{If TRUE, will breaks values into even log10 space intervals, useful when values are p-values.}

\item{include.lowest}{Whether the lowest interval should include the lowest value, passed to cut}

\item{msa}{A multiple sequence alignment, such as those returned by readBstringSet}

\item{top.name}{Name of entity, which will be set as top name and will refer to first entity in plot}

\item{sort}{Sort entities by alphabetic order}

\item{comp.seq}{Option to compare sequences from different entities by their "name" and provide
proper covariance colouring in future}

\item{counts}{An array of integers the same length as cols, dictating the number
of times each corresponding colour should be used. When missing, the function will
divide the number of helices evenly over each of the colours available.}
}
\description{
Functions to generate colours for helices by various rules,
including integer counts, value ranges, percent identity covariation,
conservation, percentage canonical basepair, basepair frequency,
and non-pseudoknotted groups.
}
