% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_mltp.R
\name{is.helix.mltp}
\alias{is.helix.mltp}
\alias{as.helix.mltp}
\title{Check a Helix Data Table}
\usage{
is.helix.mltp(helix)

as.helix.mltp(x, length = NULL)
}
\arguments{
\item{helix}{helix structure to check}
}
\description{
Functions to check whether a structure is a valid helix data table and generate helix from data.table.
}
\author{
Volodymyr Tsybulskyi
}
