% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{plotContact}
\alias{plotContact}
\title{Plots helices in diagonal arc diagram}
\usage{
plotContact(
  helix,
  direction = "se",
  top.name = "FALSE",
  sort = TRUE,
  dist.part = 0.2,
  stable = FALSE,
  line = TRUE,
  arrow = TRUE,
  col.line = "black",
  col.arrow = "black",
  col = "black",
  shape = "triangle",
  scale = FALSE,
  debug = FALSE,
  add = FALSE,
  flip = FALSE,
  arc.lty = 1,
  append = FALSE,
  ...
)
}
\arguments{
\item{helix}{Helix data.tables, with the 6 mandatory columns. "col" refer to a styling column, and will be used for styling the helix. See example for styling usage.}

\item{top.name}{Name of entity, which will be set as top name and will refer to first entity in plot}

\item{sort}{sort entities by alphabetic order}

\item{dist.part}{distance between entities, dist.part*average length of entities}

\item{stable}{used to have same height of resulted plot, based on length of entities}

\item{line}{If TRUE, a horizontal line representing the sequence is plotted.}

\item{arrow}{If TRUE, an arrow is played on the right end of the line.}

\item{col.line}{Colour of line, default is "black"}

\item{col.arrow}{Colour of arrow, default is "black"}

\item{col}{colour of arcs, if there is no style column in helix data.table sstructure}

\item{shape}{One of "circle", "triangle", or "square", specifying the shape of the arcs.}

\item{scale}{If TRUE, inserts a scale on the plot.}

\item{debug}{If TRUE, frames the boundaries of the intended plotting space in red, used to determine if inputs produce expected output area. Also outputs to STDIN dimensions of the plot.}

\item{add, append}{If TRUE, graphical elements are added to the active plot device, else a new plot device is created for the plot.}

\item{arc.lty}{define lty of arcs}
}
\description{
Plots a helix data frame as an diagonal arc diagram,
with styling possible with properly named additional columns on the data frame.
}
