% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_line.R
\name{plotDoubleHelixMltpSingleLine}
\alias{plotDoubleHelixMltpSingleLine}
\title{Plot Helix in arc diagram in single line mode with double helix}
\usage{
plotDoubleHelixMltpSingleLine(
  top,
  bottom,
  top.name = "FALSE",
  sort = TRUE,
  dist.part = 0.2,
  stable = FALSE,
  line = TRUE,
  arrow = TRUE,
  col.line = "black",
  col.arrow = "black",
  col = "black",
  shape = "circle",
  scale = FALSE,
  debug = FALSE,
  add = FALSE,
  y = 0,
  ...
)
}
\arguments{
\item{top, bottom}{helixs data.table, with the six mandatory columns. "col" column can be specified as styling column.}

\item{top.name}{Name of entity, which will be set as top name and will refer to first entity in plot}

\item{sort}{sort entities by alphabetic order}

\item{dist.part}{distance between entities, dist.part*average length of entities}

\item{stable}{used to have same height of resulted plot, based on length of entities}

\item{line}{If TRUE, a horizontal line representing the sequence is plotted.}

\item{arrow}{If TRUE, an arrow is played on the right end of the line.}

\item{col.line}{Colour of line, default is "black"}

\item{col.arrow}{Colour of arrow, default is "black"}

\item{col}{colour of arcs, if there is no style column in helix data.table sstructure}

\item{shape}{One of "circle", "triangle", or "square", specifying the shape of the arcs.}

\item{scale}{If TRUE, inserts a scale on the plot.}

\item{debug}{If TRUE, frames the boundaries of the intended plotting space in red, used to determine if inputs produce expected output area. Also outputs to STDIN dimensions of the plot.}

\item{add}{If TRUE, graphical elements are added to the active plot device, else a new plot device is created for the plot.}

\item{y}{The vertical offset of the arc base relative to 0 along the y-axis.}
}
\description{
Plots a helix data table as an arc diagram in single line, where 2 helix are given by user.
with styling possible with properly named additional columns on the data table "col",
can be used separately in each helix data.table.
}
\author{
Volodymyr Tsybulskyi
}
