% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_line.R, R/single_line.R
\name{plotHelixMltpDoubleLine}
\alias{plotHelixMltpDoubleLine}
\alias{plotHelixMltpSingleLine}
\alias{plotComparisonHelixMltpSingleLine}
\title{Plot Helix in arc diagram in single or double line mode}
\usage{
plotHelixMltpDoubleLine(
  helix,
  top.name = "FALSE",
  sort = FALSE,
  dist.between = 0.1,
  dist.part = 0.2,
  stable = FALSE,
  line = TRUE,
  arrow = TRUE,
  col.line = "black",
  col.arrow = "black",
  col = "black",
  shape = "circle",
  scale = FALSE,
  debug = FALSE,
  add = FALSE,
  y = 0,
  arc.lty = 1,
  append = FALSE,
  x = 0,
  ...
)

plotHelixMltpSingleLine(
  helix,
  top.name = "FALSE",
  sort = TRUE,
  dist.part = 0.2,
  stable = FALSE,
  line = TRUE,
  arrow = TRUE,
  col.line = "black",
  col.arrow = "black",
  col = "black",
  shape = "circle",
  scale = FALSE,
  debug = FALSE,
  add = FALSE,
  flip = FALSE,
  y = 0,
  arc.lty = 1,
  ...
)

plotComparisonHelixMltpSingleLine(
  helix1,
  helix2,
  top.name = "FALSE",
  sort = TRUE,
  dist.between = 0.1,
  dist.part = 0.2,
  stable = FALSE,
  line = TRUE,
  arrow = TRUE,
  col.line = "black",
  col.arrow = "black",
  col = "black",
  shape = "circle",
  scale = FALSE,
  debug = FALSE,
  append = TRUE,
  add = FALSE,
  cols = NA,
  breaks = NA,
  log = FALSE,
  include.lowest = TRUE,
  ...
)
}
\arguments{
\item{helix}{Helix data.tables, with the 6 mandatory columns. "col" refer to a styling column, and will be used for styling the helix. See example for styling usage.}

\item{top.name}{Name of entity, which will be set as top name and will refer to first entity in plot}

\item{sort}{sort entities by alphabetic order}

\item{dist.between}{distance between entities in y axes, dist.between*average length of entities}

\item{dist.part}{distance between entities, dist.part*average length of entities}

\item{stable}{used to have same height of resulted plot, based on length of entities}

\item{line}{If TRUE, a horizontal line representing the sequence is plotted.}

\item{arrow}{If TRUE, an arrow is played on the right end of the line.}

\item{col.line}{Colour of line, default is "black"}

\item{col.arrow}{Colour of arrow, default is "black"}

\item{col}{colour of arcs, if there is no style column in helix data.table sstructure}

\item{shape}{One of "circle", "triangle", or "square", specifying the shape of the arcs.}

\item{scale}{If TRUE, inserts a scale on the plot.}

\item{debug}{If TRUE, frames the boundaries of the intended plotting space in red, used to determine if inputs produce expected output area. Also outputs to STDIN dimensions of the plot.}

\item{add, append}{If TRUE, graphical elements are added to the active plot device, else a new plot device is created for the plot.}

\item{y}{The vertical offset of the arc base relative to 0 along the y-axis.}

\item{arc.lty}{define lty of arcs}

\item{flip}{If TRUE, flips the arcs upside down about the y-axis.}
}
\description{
Plots a helix data table as an arc diagram in single line or double line mode,
with styling possible with properly named additional columns on the data table.
}
\details{
plotHelixMltpSingleLine creates an arc diagram with all arcs on top and in single line mode,
entites goes from left to right.
plotHelixMltpDoubleLine creates an arc diagram with first entity below and other in top,
if you choose sort option then they will be sorted alphabetically. With top name you can define
which entity will in below.
}
\seealso{
colourByCountMltp, colourByValueMltp
}
\author{
Volodymyr Tsybulskyi
}
