% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_mltp.R
\name{readHelixMltp}
\alias{readHelixMltp}
\title{Read Helix File for Multiple entities}
\usage{
readHelixMltp(file)
}
\arguments{
\item{file}{file name which has helix in multiple entities}
}
\description{
Reads Helix file from source file into data.table format
}
\details{
\strong{Helix:} Files start with a header line beginning with # followed by the sequence length,
followed by a six-column tab-delimited table (with column names), where each row
corresponds to a helix in the structure. The four columns are i and j for the
left-most and right-most basepair positions respectively, the length of the
helix (converging inwards from i and j, an arbitrary value assigned to the helix,
ent.i and ent.j entity for left-most and right-most basepair positions respectively.
}
\examples{

file <- system.file("extdata", "known.helix.mltp.txt", package = "R4RNA")
helix <- readHelixMltp(file)
head(helix)

}
\author{
Volodymyr Tsybulskyi
}
